Added a Toolkit.diameter(Graph) method.
This method return the diameter of the graph (longest of the shortest
paths between all nodes and any other).

Also corrected a bug in APSP and changed some remaining floats to
doubles.

Changed some floats to doubles in Centroid and Eccentricity to follow
the changes in APSP.

+++ src/org/graphstream/algorithm/Centroid.java
@@ -99,7 +99,7 @@ public class Centroid implements Algorithm {
 
 			for (Node other : graph.getEachNode()) {
 				if (node != other) {
-					float d = info.getLengthTo(other.getId());
+					double d = info.getLengthTo(other.getId());
 
 					if (d < 0)
 						System.err
